Moved the CreateTabItem method to UiSharedService. MCDO: Fixed the Gpose tab glitch.
